requirements.txt
tesseract
pytesseract
opencv-python
pdf2image
Pillow
regex
